[#57]feat: form을 common/form-component에 CustomForm으로 내보냄
